Settings view cleanup (#429)
* change spend unconfirm fund Text

* disable user interaction on wallet log

* add copy button to copy wallet log

* enable user interaction and disable editing

* Remove Show debug messages on settings

* refactor settings

* hide delete button from settings, undo changes to main.storyboard

* undo changes to main.storyboard

* correct file paths

* fixing file arrangements

* move all settings-related files to Features/Settings

* remove settings and related vcs from main.storyboard

* cleanup settings

* show nav back button when settings is opened from splash screen

* cleanup wallet log copy function
